Edit actions from dashcards
