+++ frontend/src/metabase/writeback/components/ActionViz/Action.tsx
@@ -25,7 +25,7 @@ import LinkButton from "./LinkButton";
 import ActionForm from "./ActionForm";
 
 import { shouldShowConfirmation } from "./utils";
-import { ActionParameterMapper } from "./ActionParameterMapper";
+import { ActionParameterOptions } from "./ActionOptions";
 
 interface ActionProps extends VisualizationProps {
   dashcard: ActionDashboardCard;
